Add procon/prime.factorize

+++ procon/int.gcd.rs
@@ -1,2 +1,2 @@
 /// Integer function - GCD
-fn gcd(a: i32, b: i32) -> i32 { if b == 0 { a } else { gcd(b, a % b) } }
+fn gcd(a: i64, b: i64) -> i64 { if b == 0 { a } else { gcd(b, a % b) } }
